// src/app/models/IContactsViewModel.ts
export class ContactDetailsViewModel {
    AccountObj: AccountListViewModel;
    StageList: StageViewModel[];
    StateList: StateViewModel[];
    CountryList: CountryViewModel[];
    LeadSourceList: LeadSourceViewModel[];
    lstContactType: ContactTypeViewModel[];
    LstNotes: AllNoteViewModel[];
    lstDepartment: DepartmentViewModel[];
    AllNotesObj: AllNoteViewModel;
    listCustomfields: CustomManageMasterViewModel[];
    lstcustomVM: CustomValueMasterViewModel[];
    RoleID: number;
    FieldId: number;
    UserID: number;
    Module: string;
    Column_Id: number;
    ActualColumnName: string;
    Column_Label: string;
    Column_Type: string;
    Field_Type: string;
    Column_Description: string;
    HoverText: string;
    InputDataType: string;
    DefaultValue: string;
    DisplayPosition: number | null;
    MaxLength: number | null;
    RequiredField: boolean | null;
    MultiValued: boolean | null;
    ColumnDrpChkValues: string;
    CreatedDate: Date;
    ModifiedDate: Date;
    IsActive: boolean;
    IsShowInJoinNow: boolean;
  MailingcountryID: number;
  password:any
  Renterpassword:any
}

export class AccountListViewModel {
    ID: number;
    Name: string;
    Title: string;
    FirstName: string;
    MiddleName: string;
    LastName: string;
    Phone: string;
    Fax: string;
    Email: string;
    Mobile: string;
    EmailOptOut: string;
    AccountTypeId: number | null;
    Website: string;
    Ownership?: string;
    TotalEmployees: string;
    IndustryID: number | null;
    AnnualRevenue: string;
    Rating: string;
    LeadStatus: string;
    LeadSource: string;
    MailingAddress: string;
    MailingAddress2: string;
    Mailingcity: string;
    MailingstateID: number | null;
    Mailingzip: string;
    MailingcountryID: number | null;
    BillingAddress: string;
    BillingAddress2: string;
    Billingcity: string;
    BillingstateID: any | null;
    Billingzip: string;
    BillingcountryID: any;
    CreatedDate: string | null;
    ModifiedDate: string | null;
    CreatedBy: number | null;
    ModifiedBy: number | null;
    Description: string;
    SeachIndex: string;
    FacebookUsername: string;
    TwitterUsername: string;
    LinkedInUsername: string;
    SkypeUsername: string;
    ContactTypeID: number | null;
    CompanyID: number | null;
    CompanyName: string;
    CompanyDBA: string;
    OwnerID?: number ;
    EstimateBudget: number | null;
    CloseDate: string | null;
    StageID: number | null;
    BusinessTypeID: number | null;
    BusinessTypeName: string;
    Probability: string;
    NextStep: string;
    ExpectedRevenue: number | null;
    OppurtunitySourceID: number | null;
    OpportunitySource: string;
    ContactID: number | null;
    IsLead: string;
    ProdOpporID: number | null;
    GenooSource: string;
    GenooID: string;
    IsGenoo: string;
    LeadSourceText: string;
    BuyingStageText: string;
    ContactTypeText: string;
    PhoneExt: string;
    MailingStateText: string;
    MailingCountryText: string;
    BillingStateText: string;
    BillingCountryText: string;
    OptEmailOut: boolean;
    LeadName: string;
    CompanyExists: string;
    IsOpportunity: boolean;
    OppAmount: number | null;
    OppName: string;
    ContactName: string;
    Owner: string;
    OwnerFirstName: string;
    OwnerLastName: string;
    Stage: string;
    LSource: string;
    BuyingStage: string;
    MaillingCountry: string;
    MaillingStateName: string;
    MaillingAbbreviation: string;
    BillingCountry: string;
    BillingStateName: string;
    BillingAbbreviation: string;
    Keyword: string;
    RoleID: number;
    ColumnName: string;
    Datatype: string;
    Maxlength: number;
    APIService: string;
    InitialMailc: string;
    ContactCompanyName: string;
    OpporContactName: string;
    NotesId: number;
    OwnerShip1: string;
    Stage1: string;
    APIContactID: string;
    IsActivity: number;
    DepartmentID: number | null;
    DepartmentText: string;
    Activestatus: string;
    IsContactActive: boolean | null;
    FilePath: string;
    Radius: number;
    ZipCode: number;
    MaillingZip: string;
    RecordID: number;
    Distance: number;
    Postal: string;
    County: string;
    City1: string;
    Lat: string;
    Long: string;
    Country: string;
    Geo: string;
    Place: string;
    RecordsCount: number;
    Dob: string | null;
    PreferContact: string;
    Gender: any;
    Googleplus: string;
    UberConf: string;
    Enterprizenet: string;
    Billingcounty: string;
    Mailingcounty: string;
    RoleName: string;
    EmailAPILink: string;
    EmailAPIKey: string;
}
export class StageViewModel {
    StageId: number;
    StageName: string;
    Description: string;
    Status: string;
    AccountTypeID: number | null;
    RecordsCount: number;
    AccountName: string;
    lstAccountType: AccountTypeViewModel[];
    DuplicateStageName: string;
}
export class StateViewModel {
    ID: number;
    StateName: string;
    AbbrevIation: string;
    FullStateName: string;
}
export class LeadSourceViewModel {
    OSId: number;
    SourceName: string;
    Description: string;
    Status: string;
    RecordsCount: number;
    LeadSourceName: string;

}
export class ContactTypeViewModel {
    ID: number;
    Name: string;
    Description: string;
    Active: number;
    RecordsCount: number;
    Status: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    Createdby: number | null;
    DuplicateName: string;
}
export class AllNoteViewModel {
    NotesID: number;
    AccounttypeID: number | null;
    IsActivity: boolean | null;
    ContactId: number | null;
    CompanyID: number | null;
    OpportunityID: number | null;
    LeadID: number | null;
    ActivityId: number | null;
    Notes: string;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    CreatedBy: number | null;
    ModifiedBy: number | null;
    Module: string;
    Creator: string;
    Modifier: string;
    CreDate: string;
    CreTime: string;
    Moduleid: number;
    pagesize: number;
}
export class CustomManageMasterViewModel {
    FieldId: number;
    UserID: number;
    Module: string;
    Column_Id: number;
    ActualColumnName: string;
    Column_Label: string;
    Column_Type: string;
    Field_Type: string;
    Column_Description: string;
    HoverText: string;
    InputDataType: string;
    DefaultValue: string;
    DisplayPosition: number | null;
    MaxLength: number | null;
    RequiredField: boolean | null;
    MultiValued: boolean | null;
    ColumnDrpChkValues: string;
    CreatedDate: Date;
    ModifiedDate: Date;
    IsActive: boolean;
    IsShowInJoinNow: boolean;
    lstCustomOptions: CustomDrpChkValuesViewModel[];
    DrpValueId: number;
    drpFieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    flag: number;
    Fieldcount: number;
}
export class CustomValueMasterViewModel {
    CValueId: number;
    UserID: number;
    Module: string;
    ModuleRecordId: number | null;
    ModuleRecordIdTmp: number | null;
    MasterFieldId: number | null;
    CustomFieldvalue: string;
    CreatedDate: Date;
    ModifiedDate: Date;
    SearchIndex: string;
    Column_Id: number;
    ActualColumnName: string;
    Column_Label: string;
    Column_Type: string;
    Field_Type: string;
    Column_Description: string;
    HoverText: string;
    InputDataType: string;
    DefaultValue: string;
    MaxLength: number | null;
    RequiredField: boolean | null;
    lstCustomOptionsVal: CustomDrpChkValuesViewModel[];
    DrpValueId: number;
    drpFieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    CMMFieldId: number;
    FieldId: number | null;
}
export class DepartmentViewModel {
    DepartmentID: number;
    DepartmentName: string;
    Description: string;
    Active: number;
    CreatedDate: Date | null;
    ModifiedDate: Date | null;
    Createdby: number | null;
    Status: string;
    RecordsCount: number;
    DuplicateName: string;
}
export class CountryViewModel {
    CountryId: number;
    CountryName: string;
}
export class AccountTypeViewModel {
    ID: number;
    AccountTypeName: string;
    Description: string;
    Active: boolean | null;
    lstAccountType: AccountTypeViewModel[];
}
export class CustomDrpChkValuesViewModel {
    DrpValueId: number;
    FieldId: number;
    DrpValue: string;
    IsDefault: boolean;
    CreatedDate: Date | string;
    defaultval: string;
}